requirements.txt
timm
albumentations
numpy
tqdm
opencv-python
scipy
matplotlib
einops
addict
segmentation_models_pytorch

